// src/app/shared/enums/color.type.enum.ts
export enum ColorType {
  ROJO = 'red',
  AZUL = 'blue',
  VERDE = 'green',
  AMARILLO = 'yellow',
  NARANJA = 'orange',
  ROSA = 'pink',
  MORADO = 'purple',
  TURQUESA = 'turquoise',
  CELESTE = 'deepskyblue',
  LIMA = 'lime',
  DURAZNO = 'peachpuff',
  CORAL = 'coral',
  DORADO = 'gold',
  SALMON = 'salmon',
  LAVANDA = 'lavender'
}